Refactor(helm): extract dry-run render into dryrun.go
Move dryRunRender (used by webhook validation) and
  getKubeVersion (cluster Kubernetes version capabilities lookup) out
  of helm.go.

No behavior change. All 240 unit specs remain green.
